+++ .storybook/preview.ts
@@ -1,5 +1,3 @@
-import { withRouter } from 'storybook-addon-react-router-v6';
-
 import '~utils/yup/customMethods';
 import type { Preview } from '@storybook/react';
 import '../src/styles/main.global.css';
@@ -26,7 +24,5 @@ const preview: Preview = {
     },
   },
 };
-
-export const decorators = [withRouter];
 
 export default preview;
