+++ src/components/common/Extensions/SubNavigation/Partials/DropdownContent/DecideDropdown.tsx
@@ -1,46 +1,16 @@
 import React, { PropsWithChildren } from 'react';
-import { FormattedMessage, defineMessages } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 import Button from '~shared/Extensions/Button/Button';
 import { LEARN_MORE_ADMIN } from '~constants';
 
 import LinkItem from '../LinkItem';
+import LearnMore from '../LearnMore';
 
-import LearnMore from './LearnMore';
+import { MSG } from './consts';
 import styles from './DropdownItems.module.css';
 
-const displayName = 'Extensions.SubNavigation.DropdownItems.DecideDropdown';
-
-const MSG = defineMessages({
-  decisionsContentTitle: {
-    id: `${displayName}.decisionsContentTitle`,
-    defaultMessage: 'CREATE NEW DECISIONS',
-  },
-  createDecision: {
-    id: `${displayName}.createDecision`,
-    defaultMessage: 'Create Decision',
-  },
-  createDecisionDescription: {
-    id: `${displayName}.createDecisionDescription`,
-    defaultMessage: 'Create a new decision.',
-  },
-  simpleDiscussion: {
-    id: `${displayName}.simpleDiscussion`,
-    defaultMessage: 'Simple Discussion',
-  },
-  simpleDiscussionDescription: {
-    id: `${displayName}.simpleDiscussionDescription`,
-    defaultMessage: 'Discuss plans get advice from others.',
-  },
-  buttonText: {
-    id: `${displayName}.buttonText`,
-    defaultMessage: 'View decisions',
-  },
-  helpText: {
-    id: `${displayName}.helpText`,
-    defaultMessage: 'Need help with decisions? <a>Learn more</a>',
-  },
-});
+const displayName = 'common.Extensions.SubNavigation.Partials.DropdownContent.DecideDropdown';
 
 const DecideDropdown: React.FC<PropsWithChildren> = () => (
   <div>
@@ -53,11 +23,11 @@ const DecideDropdown: React.FC<PropsWithChildren> = () => (
         <LinkItem title={MSG.simpleDiscussion} description={MSG.simpleDiscussionDescription} />
       </ul>
       <div className={styles.buttonWrapper}>
-        <Button text={MSG.buttonText} mode="primaryOutline" />
+        <Button text={MSG.buttonTextDecide} mode="primaryOutline" />
       </div>
       <div className="mt-8 mb-4 text-sm text-gray-900 flex justify-center">
         <FormattedMessage
-          {...MSG.helpText}
+          {...MSG.helpTextDecide}
           values={{
             // eslint-disable-next-line react/no-unstable-nested-components
             a: (chunks) => <LearnMore chunks={chunks} href={LEARN_MORE_ADMIN} />,
