+++ src/components/common/Extensions/SubNavigation/Partials/LearnMore.tsx
@@ -3,14 +3,13 @@ import React, { PropsWithChildren } from 'react';
 import ExternalLink from '~shared/ExternalLink/ExternalLink';
 
 import { LearnMoreProps } from './types';
-import styles from './DropdownItems.module.css';
 
-const displayName = 'Extensions.SubNavigation.DropdownItems.LearnMore';
+const displayName = 'common.Extensions.SubNavigation.Partials.LearnMore';
 
 const LearnMore: React.FC<PropsWithChildren<LearnMoreProps>> = ({ chunks, href }) => {
   return (
     <ExternalLink href={href}>
-      <div className={styles.externalLink}>{chunks}</div>
+      <div className="font-semibold ml-1 text-gray-900 underline underline-offset-2 hover:cursor-pointer">{chunks}</div>
     </ExternalLink>
   );
 };
