+++ src/components/common/Extensions/SubNavigation/Partials/LinkItem/LinkItem.tsx
@@ -1,0 +1,29 @@
+import React, { PropsWithChildren } from 'react';
+import { FormattedMessage } from 'react-intl';
+
+// import Link from '~shared/Link/Link';
+
+import { LinkItemProps } from './types';
+import styles from './LinkItem.module.css';
+
+const displayName = 'common.Extensions.SubNavigation.Partials.LinkItem';
+
+const LinkItem: React.FC<PropsWithChildren<LinkItemProps>> = ({ title, description, statusBadge }) => {
+  return (
+    <li className={styles.itemWrapper}>
+      {/* <Link to="/"> */}
+      <div className={styles.title}>
+        <FormattedMessage {...title} />
+        {statusBadge}
+      </div>
+      <div className={styles.description}>
+        <FormattedMessage {...description} />
+      </div>
+      {/* </Link> */}
+    </li>
+  );
+};
+
+LinkItem.displayName = displayName;
+
+export default LinkItem;
