+++ src/components/common/Extensions/SubNavigation/SubNavigationItem/SubNavigationItemMobile.tsx
@@ -2,28 +2,23 @@ import React, { PropsWithChildren } from 'react';
 import clsx from 'clsx';
 
 import Button from '~shared/Extensions/Button/Button';
-
-import { SubNavigationItemMobileProps } from './types';
-import styles from './SubNavigationItemMobile.module.css';
 import Icon from '~shared/Icon/Icon';
 
-const displayName = 'Extensions.SubNavigation.SubNavigationItemMobile';
+import { SubNavigationItemProps } from './types';
+import styles from './SubNavigationItemMobile.module.css';
 
-const SubNavigationItemMobile: React.FC<PropsWithChildren<SubNavigationItemMobileProps>> = ({
+const displayName = 'common.Extensions.SubNavigation.SubNavigationItem.SubNavigationItemMobile';
+
+const SubNavigationItemMobile: React.FC<PropsWithChildren<SubNavigationItemProps>> = ({
   label,
   content,
   isOpen,
   setOpen,
-  id,
   icon,
 }) => {
   return (
     <li>
-      <Button
-        onClick={() => setOpen(id)}
-        mode="textButton"
-        className={clsx(styles.button, { [styles.activeButton]: isOpen })}
-      >
+      <Button onClick={setOpen} mode="textButton" className={clsx(styles.button, { [styles.activeButton]: isOpen })}>
         <Icon name={icon} />
         <div>{label}</div>
       </Button>
