+++ src/components/common/Extensions/SubNavigation/SubNavigationItem/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SubNavigationItem';
+export { default as SubNavigationItemMobile } from './SubNavigationItemMobile';
