+++ src/components/common/Extensions/SubNavigation/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SubNavigation';
+export { default as SubNavigationMobile } from './SubNavigationMobile';
