--- src/components/shared/Extensions/SubNavigation/SubNavigationItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SubNavigationItem';
