--- src/components/shared/Extensions/SubNavigation/SubNavigationItem/types.ts
@@ -1,10 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface SubNavigationItemProps {
-  label: ReactNode;
-  content: ReactNode;
-  isOpen: boolean;
-  setOpen: (value: React.SetStateAction<number | undefined>) => void;
-  id: number;
-  icon: string;
-}
