--- src/components/shared/Extensions/SubNavigation/SubNavigationItemMobile/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SubNavigationItemMobile';
