--- src/components/shared/Extensions/SubNavigation/SubNavigationMobile/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SubNavigationMobile';
