--- src/components/shared/Extensions/SubNavigation/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SubNavigation';
